requirements.txt
Flask>=3.0
pandas>=2.2
openpyxl>=3.1
gunicorn
python-dotenv>=1.0.0